Add 'lazy' and lazy aggregate combinators
Some schemas need to be lazy for defining recursive structures, so we
implement a 'lazy' combinator. However, for aggregate schemas like
'arrayOf', this could be inefficient due to recalculation, so we also
define custom versions of these in a separate namespace.

The lazy variants of 'recordOf' and 'classOf' perform the same amount of
computation (or more) as just using 'lazy' on the desired fields, but
they can be used for syntactic convenience in highly recursive
structures.