Add 'InjectSchema'/'injecting' for serializing DI
Sometimes, a domain type cannot be directly reconstructed from the
serialized representation, and needs some additional context that was
discarded during serialization. A major example is dependency-injecting
a reference down an object graph. This means that decoding the type is
not symmetrical, as a 'Schema' requires. We can get around this by
making a "base" domain type without the context, which we convert to and
from the representation type as usual. We then extend this schema with a
"project", to turn the original "true" domain type into the base domain
type, and an "inject", which takes the necessary context and returns a
way to construct a true domain type from a base domain type. This
extended schema is called an 'InjectSchema', and it can be constructed
via the 'injecting' combinator.